requirements20241127.txt
--find-links /usr/local/airflow/plugins/plugins_whl
--no-index

aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alembic
amqp
anyio
apache-airflow
apache-airflow-providers-amazon
apache-airflow-providers-celery
apache-airflow-providers-common-compat
apache-airflow-providers-common-io
apache-airflow-providers-common-sql
apache-airflow-providers-fab
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-jdbc
apache-airflow-providers-oracle
apache-airflow-providers-postgres
apache-airflow-providers-smtp
apache-airflow-providers-snowflake
apache-airflow-providers-sqlite
apispec
argcomplete
asgiref
asn1crypto
attrs
babel
beautifulsoup4
billiard
blinker
boto3
boto3-stubs
botocore
botocore-stubs
cachelib
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
clickclick
colorama
colorlog
ConfigUpdater
connexion
cron-descriptor
croniter
cryptography
Deprecated
dill
distlib
dnspython
docutils
email_validator
filelock
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
flower
frozenlist
fsspec
google-re2
googleapis-common-protos
greenlet
grpcio
gunicorn
h11
httpcore
httpx
humanize
idna
importlib-metadata
importlib_resources
inflection
itsdangerous
JayDeBeApi
Jinja2
jmespath
JPype1
jsonpath-ng
jsonschema
jsonschema-specifications
kombu
lazy-object-proxy
limits
linkify-it-py
lockfile
lxml
Mako
markdown-it-py
MarkupSafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
mdit-py-plugins
mdurl
methodtools
more-itertools
multidict
mypy-boto3-logs
mypy-boto3-sqs
numpy
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
oracledb
ordered-set
packaging
pandas
pathspec
pendulum
pip
platformdirs
pluggy
ply
prettytable
prison
prometheus_client
prompt_toolkit
protobuf
psutil
psycopg2
psycopg2-binary
pyarrow
PyAthena
pycparser
pycurl
Pygments
PyJWT
pyOpenSSL
python-daemon
python-dateutil
python-nvd3
python-slugify
pytz
PyYAML
redis
redshift-connector
referencing
requests
requests-toolbelt
rfc3339-validator
rich
rich-argparse
rpds-py
s3transfer
scramp
setproctitle
setuptools
six
sniffio
snowflake-connector-python
snowflake-sqlalchemy
sortedcontainers
soupsieve
SQLAlchemy
SQLAlchemy-JSONField
sqlalchemy-redshift
SQLAlchemy-Utils
sqlparse
statsd
tabulate
tenacity
termcolor
text-unidecode
time-machine
tomli
tomlkit
tornado
types-awscrt
types-s3transfer
typing_extensions
tzdata
uc-micro-py
unicodecsv
universal_pathlib
urllib3
vine
virtualenv
watchtower
wcwidth
Werkzeug
wirerope
wrapt
WTForms
yarl
zipp